Flocking in Clouds
I made two separate sketches, just so I could mess with it and not mess
up the original. Flocking may be the way to go, this is using the
example code from processing, just tweaked slightly. Loading pixels as
floating objects will be some work afterall, but it is totally doable..